requirements.txt
tensorflow
Flask
flask_cors
numpy
imageio
scipy
Pillow